requirements.txt
django==4.2.13
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
gunicorn==21.2.0
django-axes==6.3.1
django-jazzmin==2.6.0
django-ckeditor==6.7.0